requirements.txt
Django==1.9.2
djangorestframework==3.6.4
psycopg2==2.6.1
Pillow==3.1.1
ipdb==0.10.1